Fix exception handling on ARM.
libunwind causes issues in some circumstances (including on ARM) as it
causes C++ exceptions to stop functioning, they will not be caught.
This has been an issue since at least 2009 as documented at
https://bugzilla.redhat.com/show_bug.cgi?id=480412 . The fix is to link
gcc_s before libunwind, so that it's unwinder is still used for
exceptions.

The three RaiseException tests now pass for ARM.